feat: add xp to activity card, refactor xp system
