+++ src/app/(marketing)/profile/_hooks/useExperiences.tsx
@@ -1,9 +1,10 @@
+import { Schemas } from "@guildxyz/types"
 import useSWRImmutable from "swr/immutable"
 import { useProfile } from "./useProfile"
 
-export const useExperiences = ({ count = false }: { count?: boolean }) => {
+export const useExperiences = <T extends boolean>(count?: T) => {
   const { data: profile } = useProfile()
-  return useSWRImmutable(
+  return useSWRImmutable<T extends true ? number : Schemas["Experience"][]>(
     profile
       ? [`/v2/profiles/${profile.username}/experiences`, count && `count=${count}`]
           .filter(Boolean)
